offscreen-canvas/the-offscreen-canvas/offscreencanvas.commit.w.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/multipage/scripting.html#the-offscreen-2d-rendering-context">

<script id="myWorker" type="text/worker">

function test1(offscreenCanvas)
{
    var ctx = offscreenCanvas.getContext('2d');
    ctx.fillStyle = "#0f0";
    ctx.fillRect(0, 0, 10, 10);
    ctx.commit();
}

function isInvalidStateError(funcStr, ctx)
{
    try {
        eval(funcStr);
    } catch (e) {
        if (e instanceof DOMException && e.name == "InvalidStateError")
            return true;
        return false;
    }
}

function test2()
{
    var offscreenCanvas = new OffscreenCanvas(10, 10);
    var ctx = offscreenCanvas.getContext('2d');
    ctx.fillStyle = "#0f0";
    ctx.fillRect(0, 0, 10, 10);
    return isInvalidStateError("ctx.commit()", ctx);
}

self.onmessage = function(e) {
    switch (e.data.msg) {
        case 'test1':
            test1(e.data.data);
            self.postMessage('worker finished');
            break;
        case 'test2':
            self.postMessage(test2());
            break;
    }
};

</script>

<script>

function examPlaceholder(placeholder)
{
    var canvas = document.createElement('canvas');
    canvas.width = canvas.height = 10;
    var ctx = canvas.getContext('2d');
    ctx.drawImage(placeholder, 0, 0);
    _assertPixel(canvas, 5,5, 0,255,0,255, "5,5", "0,255,0,255");
}

function makeWorker(script)
{
    var blob = new Blob([script]);
    return new Worker(URL.createObjectURL(blob));
}

async_test(function(t) {
    var placeholder = document.createElement('canvas');
    placeholder.width = placeholder.height = 10;
    var offscreenCanvas = placeholder.transferControlToOffscreen();
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        examPlaceholder(placeholder);
    }));
    worker.postMessage({msg: 'test1', data: offscreenCanvas}, [offscreenCanvas]);
}, "Test that commit behaves correctly in a worker.");

async_test(function(t) {
    var worker = makeWorker(document.getElementById("myWorker").textContent);
    worker.addEventListener('message', t.step_func_done(function(msg) {
        assert_true(msg.data);
    }));
    worker.postMessage({msg: 'test2'});
}, "Test that call commit on an OffscreenCanvas that is not transferred from a HTMLCanvasElement throws exception in a worker.");

</script>
